requirements.txt
# Core
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Database & ORM
alembic==1.13.1
sqlalchemy-utils==0.41.1

# API Clients
eia-python==1.0.0
openenergy-py==0.1.0

# Processing
scikit-learn==1.3.2  # For anomaly detection
statsmodels==0.14.0

# Workflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-http==4.5.0

# FastAPI (for later)
fastapi==0.104.1
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0